Refactor, add logging
