+++ src/actiondoc/__init__.py
@@ -1,1 +1,1 @@
-from .action import ActionDoc
+from .main import ActionDoc
