+++ src/actiondoc/__main__.py
@@ -1,11 +1,22 @@
 import json
+import logging
 import os
 import sys
 
-from .action import ActionDoc
+from .main import ActionDoc
 
 # An entrypoint to generate Action documentation Markdown
 # using environment variables as arguments
+
+# Automatically enable debug when it is enabled in GitHub Actions:
+# https://docs.github.com/en/actions/monitoring-and-troubleshooting-workflows/enabling-debug-logging
+if os.environ.get("ACTIONS_RUNNER_DEBUG") == "true":
+    log_level = logging.DEBUG
+else:
+    log_level = logging.INFO
+
+logging.basicConfig(level=log_level)
+logging.getLogger("actiondoc")
 
 REQUIRED_ENV_VARS = [
     "ACTION_YAML_FILE",
@@ -30,7 +41,7 @@ def _load_env_vars():
         missing_env_vars = [k for k, v in env_vars.items() if not v]
         missing_env_vars_len = len(missing_env_vars)
         missing_env_vars_str = ", ".join(map(str, missing_env_vars))
-        sys.stderr.write(
+        logging.error(
             f"Missing environment "
             f"variable{'s' if missing_env_vars_len> 1 else ''}: "
             f"{missing_env_vars_str}"
@@ -41,6 +52,10 @@ def _load_env_vars():
 
 
 config = _load_env_vars()
+
+for k, v in config.items():
+    logging.info(f"{k,v}")
+
 action_doc = ActionDoc(action_file=config["ACTION_YAML_FILE"])
 action_doc.insert_markdown(
     include_inputs=json.loads(config["INCLUDE_INPUTS"].lower()),
